requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-dotenv
httpx
python-multipart
google-generativeai
